requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
Flask==2.2.2
Flask-HTTPAuth==4.7.0
github3.py==3.2.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.12.1
